inttodecimal

+++ JsonConversion/ProductV2.cs
@@ -6,6 +6,6 @@
         public string name { get; set; }
         public string price { get; set; }
         public int? count { get; set; }
-        public int[] size { get; set; }
+        public decimal[] size { get; set; }
     }
 }
